requirements.txt
Django==1.8.5
django-appconf==1.0.1
django-bootstrap3-sass==3.3.2.2
django-compressor==1.5
django-extensions==1.5.9
django-libsass==0.4
django-widget-tweaks==1.4.1
EasyProcess==0.1.9
greenlet==0.4.9
libsass==0.8.3
msgpack-python==0.4.6
neovim==0.0.38
Pillow==3.0.0
PyVirtualDisplay==0.1.5
six==1.9.0
virtualenv==13.1.2
wheel==0.24.0
django-password-reset==0.8.1
ipdb==0.8.1
djangorestframework==3.2.4